+++ src/main/java/com/jumayev/market_project/SERVICES/classes/CategoryService.java
@@ -19,7 +19,7 @@ public class CategoryService implements EntyService<CategoryDTO> {
     private final CategoryMapper mapper;
 
     @Override
-    public CategoryDTO getById(Integer target) {
+    public CategoryDTO getById(Long target) {
         return mapper.toDTO(repository.getReferenceById(target));
     }
 
@@ -34,14 +34,14 @@ public class CategoryService implements EntyService<CategoryDTO> {
     }
 
     @Override
-    public CategoryDTO update(Integer target, CategoryDTO info) {
+    public CategoryDTO update(Long target, CategoryDTO info) {
         Category temp = repository.getReferenceById(target);
         temp.setName(info.getName());
         return mapper.toDTO(repository.save(temp));
     }
 
     @Override
-    public void deleteById(Integer target) {
+    public void deleteById(Long target) {
         repository.deleteById(target);
     }
 }
